requirements.txt
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pycryptodome==3.23.0
python-dotenv==1.2.1
requests==2.32.5
tuya-connector-python==0.1.2
urllib3==2.6.3
websocket-client==1.9.0
Werkzeug==3.1.3